+++ packages/two_dimensional_scrollables/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
+
 ## 0.5.2
 
 * Fixes a crash in `TreeView` when it collapses to 0 rows or the last node is collapsed.